src/test/java/com/estsoft/springproject/blog/controller/BlogPageControllerTest.java
package com.estsoft.springproject.blog.controller;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.when;

import com.estsoft.springproject.blog.domain.Article;
import com.estsoft.springproject.blog.service.BlogService;
import java.util.List;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.ui.Model;

class BlogPageControllerTest {

  MockMvc mockMvc;

  @InjectMocks
  BlogPageController blogPageController;

  @Mock
  BlogService blogService;

  @Mock
  Model model;

  @BeforeEach
  void setUp() {
    MockitoAnnotations.openMocks(this);
    blogPageController = new BlogPageController(blogService);
    mockMvc = MockMvcBuilders.standaloneSetup(blogPageController).build();
  }

  @Test
  @Disabled("showArticle somehow not working")
  void testShowArticleList() {
    // Given
    List<Article> articles = List.of(new Article(), new Article());
    when(blogService.findAll()).thenReturn(articles);

    // When
    String viewName = blogPageController.showArticle(1L, model);

    // Then
    assertEquals("articleList", viewName);
  }

  @Test
  void testShowArticle() {
    // Given
    Long articleId = 1L;
    Article article = new Article();
    article.setId(articleId);
    article.setTitle("Title");
    article.setContent("Content");
    when(blogService.findBy(articleId)).thenReturn(article);

    // When
    String viewName = blogPageController.showArticle(articleId, model);

    // Then
    assertEquals("article", viewName);
  }

  @Test
  void testAddArticleWithValidId() {
    // Given
    Long articleId = 1L;
    Article article = new Article();
    article.setId(articleId);
    article.setTitle("Existing Title");
    article.setContent("Existing Content");
    when(blogService.findBy(articleId)).thenReturn(article);

    // When
    String viewName = blogPageController.addArticle(articleId, model);

    // Then
    assertEquals("newArticle", viewName);
  }
}